manifest_firstboot: log firstboot rpms
manifest.log list all packages installed in the image at build time,
but some packages (bindings and their dependencies) are installed during
the first boot only and do not figure in manifest.log.

This introduces a new manifest-firstboot.log image build log file which
lists packages to be installed at firstboot (and their dependencies which
are not installed yet) in the same format as the original manifest.log.